Revert "Keywords react"
