+++ README.md
@@ -3,7 +3,7 @@ To run our app:
 1. download the repository
 1. `cd` into the directory in your terminal
 1. type `npm install` in console
-1. type `npm start` to start the app
+1. type `nodemon` to start the app
 1. go to `http://localhost:8080` in your browser
 
 > Visit https://suprez.herokuapp.com/ to see this app live
