+++ package.json
@@ -1,42 +1,26 @@
 {
-	"name": "suprez",
-	"version": "1.0.0",
-	"description": "impulsive presentation web app",
-	"main": "server.js",
-	"scripts": {
-		"start": "webpack-dev-server --mode development",
-		"test": "echo \"Error: no test specified\" && exit 1"
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/noltron000/SPD-1-3_suprez.git"
-	},
-	"author": "",
-	"license": "ISC",
-	"bugs": {
-		"url": "https://github.com/noltron000/SPD-1-3_suprez/issues"
-	},
-	"homepage": "https://github.com/noltron000/SPD-1-3_suprez#readme",
-	"dependencies": {
-		"body-parser": "^1.18.3",
-		"express": "^4.16.4",
-		"method-override": "^3.0.0",
-		"mongoose": "^5.4.7",
-		"path": "^0.12.7",
-		"react": "^16.8.3",
-		"react-dom": "^16.8.3",
-		"react-hot-loader": "^4.7.1"
-	},
-	"devDependencies": {
-		"@babel/cli": "^7.2.3",
-		"@babel/core": "^7.3.3",
-		"@babel/preset-env": "^7.3.1",
-		"@babel/preset-react": "^7.0.0",
-		"babel-loader": "^8.0.5",
-		"css-loader": "^2.1.0",
-		"style-loader": "^0.23.1",
-		"webpack": "^4.29.5",
-		"webpack-cli": "^3.2.3",
-		"webpack-dev-server": "^3.2.0"
-	}
+  "name": "suprez",
+  "version": "1.0.0",
+  "description": "impulsive presentation web app",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/noltron000/SPD-1-3_suprez.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/noltron000/SPD-1-3_suprez/issues"
+  },
+  "homepage": "https://github.com/noltron000/SPD-1-3_suprez#readme",
+  "dependencies": {
+    "body-parser": "^1.18.3",
+    "express": "^4.16.4",
+    "method-override": "^3.0.0",
+    "mongoose": "^5.4.7",
+    "path": "^0.12.7"
+  }
 }
