--- pub/keyword_form.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-
-<head>
-	<meta charset="utf-8">
-	<title>Suprez!</title>
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<link rel="stylesheet" href="css/fonts.css">
-	<link rel="stylesheet" href="css/main.css">
-	<script src="js/main.js" async defer></script>
-</head>
-
-<body>
-	<form method="POST">
-		Keywords:<br>
-		<span class="input_forms">
-			<div class="input_area">
-				<input type="text" name="keyword" value="keyword">
-			</div>
-			<div class="input_area">
-				<input type="text" name="keyword" value="keyword">
-			</div>
-			<input type="submit" value="Submit">
-		</span>
-	</form>
-</body>
