--- src/components/app.jsx
@@ -1,14 +1,0 @@
-import React, { Component } from "react";
-// import "./App.css";
-
-class App extends Component {
-	render() {
-		return (
-			<div className="App">
-				<h1> Hello, World! </h1>
-			</div>
-		);
-	}
-}
-
-export default App;
