+++ static/index.html
@@ -1,25 +1,4 @@
-<!-- sourced from https://raw.githubusercontent.com/reactjs/reactjs.org/master/static/html/single-file-example.html -->
 <!DOCTYPE html>
-<html>
-
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-		<title>React Starter</title>
-	</head>
-
-	<body>
-		<h1><code>index.html</code> has loaded</h1>
-		<div id="root"></div>
-		<noscript>
-			You need to enable JavaScript to run this app.
-		</noscript>
-		<script src="../dist/bundle.js"></script>
-	</body>
-
-</html>
-
-<!-- <!DOCTYPE html>
 
 <head>
 	<meta charset="utf-8">
@@ -104,4 +83,4 @@
 	</div>
 </body>
 
-</html> -->
+</html>
