+++ static/presentation.html
@@ -9,5 +9,4 @@
 </head>
 
 <body>
-
 </body>
