--- webpack.config.js
@@ -1,34 +1,0 @@
-const path = require("path");
-const webpack = require("webpack");
-
-module.exports = {
-	entry: "./src/bundle.js",
-	mode: "development",
-	module: {
-		rules: [
-			{
-				test: /\.(js|jsx)$/,
-				exclude: /(node_modules|bower_components)/,
-				loader: "babel-loader",
-				options: { presets: ["@babel/env"] }
-			},
-			{
-				test: /\.css$/,
-				use: ["style-loader", "css-loader"]
-			}
-		]
-	},
-	resolve: { extensions: ["*", ".js", ".jsx"] },
-	output: {
-		path: path.resolve(__dirname, "dist/"),
-		publicPath: "/dist/",
-		filename: "bundle.js"
-	},
-	devServer: {
-		contentBase: path.join(__dirname, "pub/"),
-		port: 8080,
-		publicPath: "http://localhost:8080/dist/",
-		hotOnly: true
-	},
-	plugins: [new webpack.HotModuleReplacementPlugin()]
-};
